+++ CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.1.6] - 2022-07-12
 
 ### Added
-- Checkbox for `direct_apply` in jopPositions.
-- `benefits` in jopPositions.
-- Structured Data for `experienceRequirements`, `responsibilities`, `jobBenefits`, `jobLocation,Type` and `directApply`.
+- `direct_apply` option to job positions.
+- `benefits` to job positions.
+- Structured data for `experienceRequirements`, `responsibilities`, `jobBenefits`, `jobLocation` and `directApply`.
+
 ## [2.1.5] - 2022-06-30
 
 ### Fixed